chg: Renamed Resources
Updated the names to be clear what its for and which are grouped
together. There are moved blocks for all the resources. However,
this is still a breaking change in some cases, since this is
the second time the resources have been renamed. Also the
upgrade to latest providers has broken some parts of the S3
bucket out into their own resources for them to be tracked.

BREAKING CHANGE